Refactor types + comments
