+++ functions/src/constants.ts
@@ -1,1 +1,4 @@
+/**
+ * Région à utiliser par défaut pour les fonctions (Francfort)
+ */
 export const EU_WEST_3 = 'europe-west3'
