+++ functions/src/model/index.ts
@@ -1,0 +1,5 @@
+export * from './Group'
+export * from './Match'
+export * from './Notification'
+export * from './Team'
+export * from './User'
